requirements.txt
GitPython==3.1.0
pylint==2.4.4
pytest==5.4.1
selenium==3.141.0
